+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 
 
+## [2.2.0](https://github.com/futurestudio/hapi-request-utilities/compare/v2.1.1...v2.2.0) - 2019-10-17
+
+### Added
+- basic TypeScript declarations in `lib/index.d.ts`
+
+
+## [2.1.2](https://github.com/futurestudio/hapi-request-utilities/compare/v2.1.1...v2.1.2) - 2019-10-12
+
+### Updated
+- bump dependencies
+
+
+## [2.1.1](https://github.com/futurestudio/hapi-request-utilities/compare/v2.1.0...v2.1.1) - 2019-08-08
+
+### Fixed
+- `request.user` stores the authenticated credentials
+
+
 ## [2.1.0](https://github.com/futurestudio/hapi-request-utilities/compare/v2.0.0...v2.1.0) - 2019-07-29
 
 ### Added